add default_table
