+++ requirements.txt
@@ -1,3 +1,2 @@
 sqlalchemy
 openpyxl
-psutil
